requirements.txt
flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
pickle-mixin==1.0.2
requests==2.31.0
python-dotenv==1.0.0
pymysql==1.1.0
sqlalchemy==2.0.23
